+++ src/sequali/report_modules.py
@@ -35,8 +35,9 @@ import pygal  # type: ignore
 import pygal.style  # type: ignore
 
 from ._qc import A, C, G, N, T
-from ._qc import (AdapterCounter, DedupEstimator, NanoStats, PerTileQuality,
-                  QCMetrics, SequenceDuplication)
+from ._qc import (AdapterCounter, DedupEstimator,
+                  INSERT_SIZE_MAX_ADAPTER_STORE_SIZE, InsertSizeMetrics,
+                  NanoStats, PerTileQuality, QCMetrics, SequenceDuplication)
 from ._qc import NUMBER_OF_NUCS, NUMBER_OF_PHREDS, PHRED_MAX
 from ._version import __version__
 from .adapters import Adapter
@@ -1745,6 +1746,192 @@ class NanoStatsReport(ReportModule):
         """
 
 
+@dataclasses.dataclass
+class InsertSizeMetricsReport(ReportModule):
+
+    insert_sizes: Sequence[int]
+
+    @classmethod
+    def from_insert_size_metrics(cls, metrics: InsertSizeMetrics):
+        return cls(
+            insert_sizes=metrics.insert_sizes().tolist(),
+        )
+
+    def insert_sizes_plot(self):
+        insert_sizes_copy = list(self.insert_sizes)
+        insert_sizes_copy[0] = 0
+        plot = pygal.Bar(
+            title="Insert Sizes",
+            x_title="Insert size",
+            y_title="Number of reads",
+            style=ONE_SERIE_STYLE,
+            x_labels=list(range(len(insert_sizes_copy))),
+            x_labels_major_every=10,
+            show_minor_x_labels=False,
+            **COMMON_GRAPH_OPTIONS
+        )
+        plot.add("", insert_sizes_copy)
+        return plot
+
+    def to_html(self) -> str:
+        total_reads = sum(self.insert_sizes)
+        no_overlap_reads = self.insert_sizes[0]
+        return f"""
+            {html_header("Insert Sizes", 1)}
+            <p class="explanation">
+            Insert sizes are calculated by taking the first and last
+            16&#8239;bp from read 2. These are searched for in read 1 while
+            allowing at most 1 error. This method is very fast and reasonably
+            accurate, but does not account for long poly-G cycles.
+            </p>
+            <table>
+            <tr>
+                <td>Total reads</td>
+                <td style="text-align:right;">{total_reads:,}</td>
+            </tr>
+            <tr>
+                <td>Reads without overlap</td>
+                <td style="text-align:right;">{no_overlap_reads:,}</td>
+                <td style="text-align:right;">
+                    {no_overlap_reads / max(total_reads, 1):.2%}
+                </td>
+            </tr>
+            </table>
+            {figurize_plot(self.insert_sizes_plot())}
+        """
+
+
+@dataclasses.dataclass
+class AdapterFromOverlapReport(ReportModule):
+    total_reads: int
+    number_of_adapters_read1: int
+    number_of_adapters_read2: int
+    adapters_read1: Sequence[Tuple[str, int]]
+    adapters_read2: Sequence[Tuple[str, int]]
+    longest_adapter_read1: str
+    longest_adapter_read2: str
+    longest_adapter_read1_match: str
+    longest_adapter_read2_match: str
+
+    @staticmethod
+    def select_relevant_adapters(adapter_list: Sequence[Tuple[str, int]]):
+        """
+        Get the most prevalent adapters for each length. Sort the resulting
+        list on length.
+        """
+        # Sort list on count
+        sorted_list = sorted(adapter_list, reverse=True, key=lambda x: (x[1]))
+        new_list = []
+        lengths_to_get = set(range(1, INSERT_SIZE_MAX_ADAPTER_STORE_SIZE + 1))
+        for adapter, count in sorted_list:
+            if len(adapter) in lengths_to_get:
+                lengths_to_get.remove(len(adapter))
+                new_list.append((adapter, count))
+        new_list.sort(key=lambda x: len(x[0]))
+        return new_list
+
+    @classmethod
+    def from_insert_size_metrics(cls, metrics: InsertSizeMetrics):
+        def contaminant_iterator():
+            for file in DEFAULT_CONTAMINANTS_FILES:
+                yield from fasta_parser(file)
+
+        sequence_index = create_sequence_index(contaminant_iterator(),
+                                               DEFAULT_K)
+        adapters_read1 = AdapterFromOverlapReport.select_relevant_adapters(
+                metrics.adapters_read1())
+        adapters_read2 = AdapterFromOverlapReport.select_relevant_adapters(
+            metrics.adapters_read2())
+        longest_adapter_read1 = adapters_read1[-1][0]
+        longest_adapter_read2 = adapters_read2[-1][0]
+        longest_adapter_read1_match = identify_sequence(
+            longest_adapter_read1, sequence_index)[2]
+        longest_adapter_read2_match = identify_sequence(
+            longest_adapter_read2, sequence_index)[2]
+        return cls(
+            total_reads=metrics.total_reads,
+            number_of_adapters_read1=metrics.number_of_adapters_read1,
+            number_of_adapters_read2=metrics.number_of_adapters_read2,
+            adapters_read1=adapters_read1,
+            adapters_read2=adapters_read2,
+            longest_adapter_read1=longest_adapter_read1,
+            longest_adapter_read2=longest_adapter_read2,
+            longest_adapter_read1_match=longest_adapter_read1_match,
+            longest_adapter_read2_match=longest_adapter_read2_match,
+        )
+
+    def to_html(self) -> str:
+        report = io.StringIO()
+        report.write(f"""
+            {html_header("Adapter Content", 1)}
+            <p class="explanation">
+                By calculating the overlap between reads it is possible to
+                determine where the adapter is.
+            </p>
+            <table>
+            <tr>
+                <td>Total reads</td>
+                <td style="text-align:right;">{self.total_reads:,}</td>
+                <td style="text-align:right;">
+                    {self.total_reads / max(self.total_reads, 1):.2%}
+                </td>
+            </tr>
+            <tr>
+                <td>Adapters in read 1</td>
+                <td style="text-align:right;">{self.number_of_adapters_read1:,}</td>
+                <td style="text-align:right;">
+                    {self.number_of_adapters_read1 / max(self.total_reads, 1):.2%}
+                </td>
+            </tr>
+            <tr>
+                <td>Adapters in read 2</td>
+                <td style="text-align:right;">{self.number_of_adapters_read2:,}</td>
+                <td style="text-align:right;">
+                    {self.number_of_adapters_read2 / max(self.total_reads, 1):.2%}
+                </td>
+            </tr>
+            </table>
+            <table><tr><th>Longest most frequent adapter</th>
+            <th>Adapter Sequence</th><th>Best match</th></tr>
+            <tr>
+                <td>Read 1</td>
+                <td>{self.longest_adapter_read1}</td>
+                <td>{self.longest_adapter_read1_match}</td>
+            </tr>
+            <tr>
+                <td>Read 2</td>
+                <td>{self.longest_adapter_read2}</td>
+                <td>{self.longest_adapter_read2_match}</td>
+            </tr>
+            </table>
+        """)
+
+        report.write(html_header("Adapters read 1", 2))
+        report.write("<table>")
+        report.write("<tr><th>Adapter</th><th>Count</th></tr>")
+        for adapter, count in self.adapters_read1:
+            report.write(
+                f"""<tr>
+                        <td>{adapter}</td>
+                        <td style="text-align:right;">{count}</td>
+                    </tr>
+                """)
+        report.write("</table>")
+
+        report.write(html_header("Adapters read 2", 2))
+        report.write("<table>")
+        report.write("<tr><th>Adapter</th><th>Count</th></tr>")
+        for adapter, count in self.adapters_read2:
+            report.write(
+                f"""<tr>
+                        <td>{adapter}</td>
+                        <td style="text-align:right;">{count}</td>
+                    </tr>
+                """)
+        report.write("</table>")
+        return report.getvalue()
+
+
 NAME_TO_CLASS: Dict[str, Type[ReportModule]] = {
     "meta": Meta,
     "summary": Summary,
@@ -1760,6 +1947,8 @@ NAME_TO_CLASS: Dict[str, Type[ReportModule]] = {
     "duplication_fractions": DuplicationCounts,
     "overrepresented_sequences": OverRepresentedSequences,
     "nanopore_metrics": NanoStatsReport,
+    "adapter_content_from_insert_sizes": AdapterFromOverlapReport,
+    "insert_size_metrics": InsertSizeMetricsReport,
 }
 
 CLASS_TO_NAME: Dict[Type[ReportModule], str] = {
@@ -1902,15 +2091,15 @@ def qc_metrics_modules(metrics: QCMetrics,
 def calculate_stats(
         filename: str,
         metrics: QCMetrics,
-        adapter_counter: AdapterCounter,
         per_tile_quality: PerTileQuality,
         sequence_duplication: SequenceDuplication,
         dedup_estimator: DedupEstimator,
         nanostats: NanoStats,
         adapters: List[Adapter],
+        adapter_counter: Optional[AdapterCounter] = None,
         filename_reverse: Optional[str] = None,
+        insert_size_metrics: Optional[InsertSizeMetrics] = None,
         metrics_reverse: Optional[QCMetrics] = None,
-        adapter_counter_reverse: Optional[AdapterCounter] = None,
         per_tile_quality_reverse: Optional[PerTileQuality] = None,
         sequence_duplication_reverse: Optional[SequenceDuplication] = None,
         graph_resolution: int = 200,
@@ -1926,38 +2115,54 @@ def calculate_stats(
         data_ranges = list(equidistant_ranges(max_length, graph_resolution))
     modules = [
         Meta.from_filepath(filename, filename_reverse),
-        *qc_metrics_modules(metrics, data_ranges, read_pair_info=read_pair_info1),
-        AdapterContent.from_adapter_counter_adapters_and_ranges(
-            adapter_counter, adapters, data_ranges, read_pair_info=read_pair_info1),
-        PerTileQualityReport.from_per_tile_quality_and_ranges(
-            per_tile_quality, data_ranges, read_pair_info=read_pair_info1),
-        DuplicationCounts.from_dedup_estimator(dedup_estimator),
+    ]
+    # Generic modules for both read1 and read2 come first.
+    if insert_size_metrics:
+        modules.append(
+            AdapterFromOverlapReport.from_insert_size_metrics(insert_size_metrics))
+        modules.append(
+            InsertSizeMetricsReport.from_insert_size_metrics(insert_size_metrics))
+    if filename_reverse:
+        modules.append(
+            DuplicationCounts.from_dedup_estimator(dedup_estimator),
+        )
+    modules.extend(qc_metrics_modules(metrics, data_ranges,
+                                      read_pair_info=read_pair_info1))
+    if adapter_counter:
+        modules.append(
+            AdapterContent.from_adapter_counter_adapters_and_ranges(
+                adapter_counter, adapters, data_ranges, read_pair_info=read_pair_info1)
+        )
+    modules.append(PerTileQualityReport.from_per_tile_quality_and_ranges(
+        per_tile_quality, data_ranges, read_pair_info=read_pair_info1),)
+    if not filename_reverse:
+        modules.append(
+            DuplicationCounts.from_dedup_estimator(dedup_estimator)
+        )
+    modules.append(
         OverRepresentedSequences.from_sequence_duplication(
             sequence_duplication,
             fraction_threshold=fraction_threshold,
             min_threshold=min_threshold,
             max_threshold=max_threshold,
             read_pair_info=read_pair_info1,
-        ),
-        NanoStatsReport.from_nanostats(nanostats)
-    ]
-    if (metrics_reverse and adapter_counter_reverse
-            and per_tile_quality_reverse and sequence_duplication_reverse):
+        )
+    )
+
+    if (metrics_reverse and per_tile_quality_reverse and
+            sequence_duplication_reverse):
         max_length_reverse = metrics_reverse.max_length
         if max_length_reverse > 500:
             data_ranges_reverse = list(logarithmic_ranges(max_length_reverse))
         else:
             data_ranges_reverse = list(
                 equidistant_ranges(max_length_reverse, graph_resolution))
+
         modules.extend(qc_metrics_modules(metrics_reverse, data_ranges_reverse,
                                           read_pair_info=READ2))
-        modules.append(AdapterContent.from_adapter_counter_adapters_and_ranges(
-            adapter_counter_reverse, adapters, data_ranges_reverse,
-            read_pair_info=READ2,
-        ))
-        PerTileQualityReport.from_per_tile_quality_and_ranges(
+        modules.append(PerTileQualityReport.from_per_tile_quality_and_ranges(
             per_tile_quality_reverse, data_ranges_reverse,
-            read_pair_info=READ2),
+            read_pair_info=READ2))
         modules.append(OverRepresentedSequences.from_sequence_duplication(
             sequence_duplication_reverse,
             fraction_threshold=fraction_threshold,
@@ -1965,4 +2170,6 @@ def calculate_stats(
             max_threshold=max_threshold,
             read_pair_info=READ2,
         ))
+
+    modules.append(NanoStatsReport.from_nanostats(nanostats))
     return modules