+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ Changelog
 
 version 0.8.0-dev
 -----------------
++ Calculate insert sizes and used adapters based on overlap between the
+  read pairs.
 + Both reads from paired-end reads are taken into consideration when
   evaluating the duplication rate.
 + Support for paired-end reads added.